requirements.txt
streamlit==1.32.0
pandas==2.2.2
openai==1.30.1
langchain==0.0.284
python-dotenv==1.0.1
tiktoken==0.4.0